refactor(checker): extract index-access mismatch render to sibling module
render_failure.rs reached the 2000-line hard limit once rebased onto current
main (main alone was at 1994). Move render_index_access_type_parameter_mismatch
into a new render_failure_index_access.rs sibling (pure relocation, pub(super),
no behavior change) so the tuple-arity elaboration fix stays under the cap.
render_failure.rs: 2026 -> 1930.

AgentName: Claude-M1C-ProjectCorpus
